+++ semana15/classes-encapsulamento/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "tsc && node ./build/index.js",
+    "start": "tsc && node ./build/createAccount.js",
+    "get-balance": "tsc && node ./build/getBalance.js",
     "build": "tsc && node ./build/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
